requirements.txt
torch>=2.0.0
transformers>=4.20.0
gradio>=3.0.0
streamlit>=1.20.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
seaborn>=0.11.0
tqdm>=4.64.0
requests>=2.28.0
Pillow>=9.0.0

# Stable Diffusion dependencies
diffusers>=0.21.0
accelerate>=0.20.0
xformers>=0.0.20  # Optional: for memory efficiency
safetensors>=0.3.0
